requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
python-dotenv
typesense
langchain_openai
langgraph
langchain-google-genai
numpy<2